Combined data store setter and getter traits into one aware trait
