--- test/unit/SetDataStoreCapableTraitTest.php
@@ -1,227 +1,0 @@
-<?php
-
-namespace Dhii\Data\Object\UnitTest;
-
-use ArrayObject;
-use Dhii\Data\Object\SetDataStoreCapableTrait as TestSubject;
-use InvalidArgumentException;
-use Xpmock\TestCase;
-use Exception as RootException;
-use PHPUnit_Framework_MockObject_MockObject as MockObject;
-
-/**
- * Tests {@see TestSubject}.
- *
- * @since [*next-version*]
- */
-class SetDataStoreCapableTraitTest extends TestCase
-{
-    /**
-     * The class name of the test subject.
-     *
-     * @since [*next-version*]
-     */
-    const TEST_SUBJECT_CLASSNAME = 'Dhii\Data\Object\SetDataStoreCapableTrait';
-
-    /**
-     * Creates a new instance of the test subject.
-     *
-     * @since [*next-version*]
-     *
-     * @param array $methods The methods to mock.
-     *
-     * @return MockObject The new instance.
-     */
-    public function createInstance($methods = [])
-    {
-        $methods = $this->mergeValues($methods, [
-            '__',
-        ]);
-
-        $mock = $this->getMockBuilder(static::TEST_SUBJECT_CLASSNAME)
-            ->setMethods($methods)
-            ->getMockForTrait();
-
-        $mock->method('__')
-                ->will($this->returnArgument(0));
-
-        return $mock;
-    }
-
-    /**
-     * Merges the values of two arrays.
-     *
-     * The resulting product will be a numeric array where the values of both inputs are present, without duplicates.
-     *
-     * @since [*next-version*]
-     *
-     * @param array $destination The base array.
-     * @param array $source      The array with more keys.
-     *
-     * @return array The array which contains unique values
-     */
-    public function mergeValues($destination, $source)
-    {
-        return array_keys(array_merge(array_flip($destination), array_flip($source)));
-    }
-
-    /**
-     * Creates a mock that both extends a class and implements interfaces.
-     *
-     * This is particularly useful for cases where the mock is based on an
-     * internal class, such as in the case with exceptions. Helps to avoid
-     * writing hard-coded stubs.
-     *
-     * @since [*next-version*]
-     *
-     * @param string $className      Name of the class for the mock to extend.
-     * @param string $interfaceNames Names of the interfaces for the mock to implement.
-     *
-     * @return object The object that extends and implements the specified class and interfaces.
-     */
-    public function mockClassAndInterfaces($className, $interfaceNames = [])
-    {
-        $paddingClassName = uniqid($className);
-        $definition = vsprintf('abstract class %1$s extends %2$s implements %3$s {}', [
-            $paddingClassName,
-            $className,
-            implode(', ', $interfaceNames),
-        ]);
-        eval($definition);
-
-        return $this->getMockForAbstractClass($paddingClassName);
-    }
-
-    /**
-     * Creates a new Invalid Argument exception.
-     *
-     * @since [*next-version*]
-     *
-     * @param string $message The exception message.
-     *
-     * @return InvalidArgumentException The new exception.
-     */
-    public function createInvalidArgumentException($message = '')
-    {
-        $mock = $this->getMockBuilder('InvalidArgumentException')
-            ->setConstructorArgs([$message])
-            ->getMock();
-
-        return $mock;
-    }
-
-    /**
-     * Creates a new exception.
-     *
-     * @since [*next-version*]
-     *
-     * @param string $message The exception message.
-     *
-     * @return RootException The new exception.
-     */
-    public function createException($message = '')
-    {
-        $mock = $this->getMockBuilder('Exception')
-            ->setConstructorArgs([$message])
-            ->getMock();
-
-        return $mock;
-    }
-
-    /**
-     * Creates a new store mock.
-     *
-     * @since [*next-version*]
-     *
-     * @return ArrayObject|MockObject The new store mock.
-     */
-    public function createStore($data = [], $methods = [])
-    {
-        $methods = $this->mergeValues($methods, [
-        ]);
-
-        $mock = $this->getMockBuilder('ArrayObject')
-            ->setMethods($methods)
-            ->getMock($data);
-
-        return $mock;
-    }
-
-    /**
-     * Tests whether a valid instance of the test subject can be created.
-     *
-     * @since [*next-version*]
-     */
-    public function testCanBeCreated()
-    {
-        $subject = $this->createInstance();
-
-        $this->assertInternalType(
-            'object',
-            $subject,
-            'A valid instance of the test subject could not be created.'
-        );
-    }
-
-    /**
-     * Tests that `_setDataStore()` works as expected when given an {@see ArrayAccess}.
-     *
-     * @since [*next-version*]
-     */
-    public function testSetDataStoreArrayAccess()
-    {
-        $store = $this->createStore();
-        $subject = $this->createInstance();
-        $_subject = $this->reflect($subject);
-
-        $initialState = $_subject->dataStore;
-        $this->assertNull($initialState, 'The initial state of the subject is wrong');
-
-        $_subject->_setDataStore($store);
-        $modifiedState = $_subject->dataStore;
-        $this->assertSame($store, $modifiedState, 'The modified state of the subject is wrong');
-    }
-
-    /**
-     * Tests that `_setDataStore()` works as expected when given null.
-     *
-     * @since [*next-version*]
-     */
-    public function testSetDataStoreNull()
-    {
-        $store = null;
-        $subject = $this->createInstance();
-        $_subject = $this->reflect($subject);
-
-        $_subject->dataStore = uniqid('store');
-        $_subject->_setDataStore($store);
-        $modifiedState = $_subject->dataStore;
-        $this->assertSame($store, $modifiedState, 'The modified state of the subject is wrong');
-    }
-
-    /**
-     * Tests that `_setDataStore()` fails as expected when given an invalid store.
-     *
-     * @since [*next-version*]
-     */
-    public function testSetDataStoreFailureInvalidStore()
-    {
-        $exception = $this->createInvalidArgumentException('Invalid store');
-        $store = uniqid('store');
-        $subject = $this->createInstance();
-        $_subject = $this->reflect($subject);
-
-        $subject->expects($this->exactly(1))
-                ->method('_createInvalidArgumentException')
-                ->with(
-                    $this->isType('string'),
-                    null,
-                    null,
-                    $store
-                )
-                ->will($this->returnValue($exception));
-
-        $this->setExpectedException('InvalidArgumentException');
-        $_subject->_setDataStore($store);
-    }
-}
